Continued on front-end work, added additional functionality
Added additional stats to the website, added being able to register accounts, logout, and forgot password from the front-end, also added the linking and unlinking of LoL accounts to LS accounts to front-end. Began to add recommendation resources to the front-end. started to try and add more niche LoL account information but does not work yet.